Backend/pyproject.toml
[tool.poetry]
name = "ripple"
version = "0.1.0"
description = "portfolio project"
authors = ["VoiceOfDarkness <abil.samedov2016@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
fastapi = {extras = ["uvicorn"], version = "^0.111.0"}
sqlmodel = "^0.0.18"
alembic = "^1.13.1"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
dependency-injector = "^4.41.0"
psycopg2-binary = "^2.9.9"
bcrypt = "^4.1.3"
passlib = "^1.7.4"
python-jose = "^3.3.0"
icecream = "^2.1.3"
authlib = "^1.3.0"
itsdangerous = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
